Rename: 유효값 검사하는 클래스를 src밑으로 이동
- 프로그래밍 요구사항에
"프로그래밍 요구 사항에서 달리 명시하지 않는 한 파일, 패키지 이름을 수정하거나 이동하지 않는다." 라고 있음
- Lotto.js 파일을 원래 위치로 이동하기 위함.
(+) 로또 정답 리스트 입력 기능 다듬음
