+++ docs/README.md
@@ -29,11 +29,13 @@
 
 - [x] 각 로또마다 몇 개의 수가 맞았는지 계산한다.
     - [x] 당첨 통계 출력
-        - 3개 일치 (5,000원)
-        - 4개 일치 (50,000원)
-        - 5개 일치 (1,500,000원)
-        - 5개 일치, 보너스 볼 일치 (30,000,000원)
-        - 6개 일치 (2,000,000,000원)
+    ```
+    - 3개 일치 (5,000원)
+    - 4개 일치 (50,000원)
+    - 5개 일치 (1,500,000원)
+    - 5개 일치, 보너스 볼 일치 (30,000,000원)
+    - 6개 일치 (2,000,000,000원)
+    ```
 
 - [x] 수익률 계산기능
     - [x] 수익률 출력 (소수점 둘째 자리에서 반올림)
@@ -43,10 +45,6 @@
 ## :file_folder: 파일 구조
 ```
 📦src
- ┣ 📂check-available
- ┃ ┣ 📜Bonus.js
- ┃ ┣ 📜Lotto.js
- ┃ ┗ 📜Price.js
  ┣ 📂lotto-make-and-calculate
  ┃ ┣ 📜Calculate.js
  ┃ ┣ 📜MakeLottos.js
@@ -56,7 +54,10 @@
  ┣ 📂utils
  ┃ ┣ 📜Constants.js
  ┃ ┗ 📜LottoHandler.js
- ┗ 📜App.js
+ ┣ 📜App.js
+ ┣ 📜Bonus.js
+ ┣ 📜Lotto.js
+ ┗ 📜Price.js
  ```
 
  <br />
@@ -70,7 +71,7 @@
  - 프로세스 실행
  - 값 입력, 로또값 생성, 계산 등 프로세스를 순서대로 실행할 수 있게 함
 
- ### check-available
+ ### Lotto.js, Bonus.js, Price.js
  - 입력받은 값에 대한 유효성을 검사
  - 예외상황일 경우 throw문으로 예외처리
